CRUD Rating + Slider

+++ ShoppingCard/Areas/Admin/Controllers/ProductController.cs
@@ -117,7 +117,7 @@ namespace ShoppingCard.Areas.Controllers
             return View(product);
         }
 
-        public async Task<IActionResult> Edit(int Id)
+        public async Task<IActionResult> Edit(long Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
             ViewBag.Categories = new SelectList(_dataContext.Categories, "Id", "Name", product.CategoryId);
@@ -210,7 +210,7 @@ namespace ShoppingCard.Areas.Controllers
         }
 
         //-- Delete product --
-        public async Task<IActionResult> Delete(int Id)
+        public async Task<IActionResult> Delete(long Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
